requirements.txt
openai>=2.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
Pillow>=10.0.0
pandas>=1.5.0
pypdf>=3.0.0
